:sparkles: Fixed detection of voice interactions

+++ src/utils/games/textInteraction.ts
@@ -9,29 +9,25 @@ export default async function (message: Message) {
   });
 
   if (interactionData) {
-    console.log(`Interation Step 1`);
-    if (interactionData.speakingTimes > textInteractionsConfig.textStreak) {
-      console.log(`Interation Completed`);
+    const lastMessage = (await message.channel.messages.fetch()).array()[1];
+    if (
+      interactionData.speakingTimes &&
+      interactionData.speakingTimes > textInteractionsConfig.textStreak
+    ) {
       await givePoints(message.author, "textInteraction");
       await interactionData.deleteOne();
     } else if (
-      !message.channel.lastMessage.author.bot &&
-      message.channel.lastMessage.author !== message.author &&
-      interactionData.lastSpeakingTime - Date.now() >
-        textInteractionsConfig.resetInterval
+      !lastMessage.author.bot &&
+      lastMessage.author !== message.author &&
+      (!interactionData.lastSpeakingTime ||
+        Date.now() - interactionData.lastSpeakingTime >
+          textInteractionsConfig.resetInterval)
     ) {
-      console.log(`Interation Step 3`);
-      await TextInteractionModel.updateOne(
-        {
-          userId: message.author.id,
-        },
-        {
-          $inc: { userSpeakingTimes: 1 },
-        }
-      );
+      interactionData.speakingTimes += 1;
+      interactionData.lastSpeakingTime = Date.now();
+      await interactionData.save();
     }
   } else {
-    console.log(`Interation Step 2`);
     await TextInteractionModel.create({
       userId: message.author.id,
       speakingTimes: 0,
